Сделал разделение рюкзака на разделы Чуть чуть начал разгребать messagehandler
