+++ handlers/messageHandler.go
@@ -32,14 +32,13 @@ func messageResolver(update tg.Update) tg.MessageConfig {
 }
 
 func callBackResolver(update tg.Update) (tg.MessageConfig, bool) {
-	msg.ChatID = update.CallbackQuery.Message.Chat.ID
 	buttons := tg.ReplyKeyboardMarkup{}
 	charData := strings.Fields(update.CallbackQuery.Data)
 	deletePrevMessage := true
 
 	userTgId := r.GetUserTgId(update)
 	user := r.GetUser(r.User{TgId: userTgId})
-	ItemLeftHand, ItemRightHand, ItemHead := usersHandItemsView(user)
+	ItemLeftHand, ItemRightHand, ItemHead := helpers.UsersHandItemsView(user)
 
 	if len(charData) == 1 {
 		switch charData[0] {
@@ -50,80 +49,100 @@ func callBackResolver(update tg.Update) (tg.MessageConfig, bool) {
 		}
 	}
 
+	fmt.Println(charData)
+
 	switch charData[0] {
+
+	// Действия в рюкзаке
+	case v.GetString("callback_char.category"):
+		resUserItems := r.GetBackpackItems(user.ID, charData[1])
+		msg.Text = helpers.MessageBackpackUserItems(resUserItems, 0, charData[1])
+		msg.ReplyMarkup = helpers.BackpackInlineKeyboard(resUserItems, 0, charData[1])
 	case v.GetString("callback_char.backpack_moving"):
-		msg = BackPackMoving(charData, update)
-	case v.GetString("callback_char.goods_moving"):
-		msg = GoodsMoving(charData, update)
+		msg = helpers.BackPackMoving(charData, user)
 	case v.GetString("callback_char.eat_food"):
-		UserEatItem(update, charData)
-	case v.GetString("callback_char.delete_item"):
-		UserDeleteItem(update, charData)
+		msg = helpers.UserEatItem(update, charData)
+
+	// Действия в инвентаре
+	case v.GetString("callback_char.goods_moving"):
+		msg = helpers.GoodsMoving(charData, update)
 	case v.GetString("callback_char.dress_good"):
-		msg = dressUserItem(update, charData)
-	case v.GetString("callback_char.take_off_good"):
-		userTakeOffGood(update, charData)
+		msg = helpers.DressUserItem(update, charData)
 	case v.GetString("callback_char.change_left_hand"), v.GetString("callback_char.change_right_hand"):
 		userItem := r.UserItem{ID: r.ToInt(charData[1])}.UserGetUserItem()
 		updateUserHand(update, charData, userItem)
 		charDataForOpenGoods := strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.goods_moving"), charData[2]))
-		msg = GoodsMoving(charDataForOpenGoods, update)
+		msg = helpers.GoodsMoving(charDataForOpenGoods, update)
 		msg.Text = fmt.Sprintf("%s%sВы надели %s", msg.Text, v.GetString("msg_separator"), userItem.Item.View)
+	case v.GetString("callback_char.take_off_good"):
+		msg = helpers.UserTakeOffGood(update, charData)
+
+	// Удаление, выкидывание, описание итема
+	case v.GetString("callback_char.delete_item"):
+		msg = helpers.UserDeleteItem(update, charData)
+	case v.GetString("callback_char.throw_out_item"):
+		msg = helpers.UserWantsToThrowOutItem(update, user, charData)
+	case v.GetString("callback_char.count_of_delete"):
+		msg = helpers.UserThrowOutItem(update, user, charData)
+	case v.GetString("callback_char.description"):
+		msg.Text = r.UserItem{ID: r.ToInt(charData[1])}.GetFullDescriptionOfUserItem()
+		msg.ReplyMarkup = helpers.DescriptionInlineButton(charData)
+
+	// Профиль
 	case v.GetString("callback_char.change_avatar"):
 		res := r.User{Avatar: charData[1]}.UpdateUser(update)
 		msg.Text = r.GetUserInfo(update)
 		msg.ReplyMarkup = helpers.ProfileKeyboard(res)
-	case v.GetString("callback_char.description"):
-		msg.Text = r.UserItem{ID: r.ToInt(charData[1])}.GetFullDescriptionOfUserItem()
-		deletePrevMessage = false
+
+	// Крафтинг
 	case v.GetString("callback_char.workbench"):
-		msg = Workbench(nil, charData)
+		msg = helpers.Workbench(nil, charData)
 	case v.GetString("callback_char.receipt"):
-		msg.Text = fmt.Sprintf("📖 *Рецепты*: 📖%s%s", v.GetString("msg_separator"), AllReceiptsMsg())
+		msg.Text = fmt.Sprintf("📖 *Рецепты*: 📖%s%s", v.GetString("msg_separator"), helpers.AllReceiptsMsg())
 		msg.ReplyMarkup = nil
 		deletePrevMessage = false
 	case v.GetString("callback_char.put_item"):
-		userItem := r.GetUserItems(user.ID)
+		userItem := r.GetUserItemsByType(user.ID, strings.Fields("food resource"))
 		msg.ReplyMarkup = helpers.ChooseUserItemButton(userItem, charData)
-		msg = OpenWorkbenchMessage(charData)
-		msg.Text = fmt.Sprintf("%s%sВыбери предмет:", msg.Text, v.GetString("msg_separator"))
+		msg.Text = fmt.Sprintf("%s%sВыбери предмет:", helpers.OpenWorkbenchMessage(charData), v.GetString("msg_separator"))
 	case v.GetString("callback_char.put_count_item"):
-		msg = OpenWorkbenchMessage(charData)
-		msg = PutCountComponent(charData)
-		msg.Text = fmt.Sprintf("%s%s⚠️ Сколько выкладываешь?", msg.Text, v.GetString("msg_separator"))
+		msg = helpers.PutCountComponent(charData)
+		msg.Text = fmt.Sprintf("%s%s⚠️ Сколько выкладываешь?", helpers.OpenWorkbenchMessage(charData), v.GetString("msg_separator"))
 	case v.GetString("callback_char.make_new_item"):
-		resp := GetReceiptFromData(charData)
+		resp := helpers.GetReceiptFromData(charData)
 		receipt := r.FindReceiptForUser(resp)
-		msg, deletePrevMessage = UserCraftItem(user, receipt)
+		msg, deletePrevMessage = helpers.UserCraftItem(user, receipt)
+
+	// Использование надетых итемов
 	case v.GetString("message.emoji.hand"), ItemLeftHand.View, ItemRightHand.View:
-		msg = UserUseHandOrInstrument(update, charData)
+		msg = helpers.UserUseHandOrInstrument(update, charData)
 	case v.GetString("message.emoji.foot"):
-		msg = UserMoving(update, user, charData[1])
+		msg = helpers.UserMoving(update, user, charData[1])
 	case ItemHead.View:
-		res := directionCell(update, charData[1])
+		res := helpers.DirectionCell(update, charData[1])
 		text, err := r.UpdateUserInstrument(update, user, ItemHead)
 		if err != nil {
 			msg.Text = fmt.Sprintf("%s%s%s", r.ViewItemInfo(res), v.GetString("msg_separator"), text)
 		} else {
 			msg.Text = r.ViewItemInfo(res)
 		}
-	case v.GetString("callback_char.throw_out_item"):
-		userWantsToThrowOutItem(update, charData)
-	case v.GetString("callback_char.count_of_delete"):
-		msg = userThrowOutItem(update, user, charData)
+
+	// Квесты
 	case v.GetString("callback_char.quests"):
 		msg.Text = v.GetString("user_location.tasks_menu_message")
 		msg.ReplyMarkup = helpers.AllQuestsMessageKeyboard(user)
 	case v.GetString("callback_char.quest"):
-		msg = OpenQuest(uint(r.ToInt(charData[1])), user)
+		msg = helpers.OpenQuest(uint(r.ToInt(charData[1])), user)
 	case v.GetString("callback_char.user_get_quest"):
 		r.UserQuest{
 			UserId:  user.ID,
 			QuestId: uint(r.ToInt(charData[1])),
 		}.GetOrCreateUserQuest()
-		msg = OpenQuest(uint(r.ToInt(charData[1])), user)
+		msg = helpers.OpenQuest(uint(r.ToInt(charData[1])), user)
 	case v.GetString("callback_char.user_done_quest"):
-		msg = UserDoneQuest(uint(r.ToInt(charData[1])), user)
+		msg = helpers.UserDoneQuest(uint(r.ToInt(charData[1])), user)
+
+	// Дом юзера
 	case v.GetString("callback_char.buy_home"):
 		err := user.CreateUserHouse()
 		text := "Поздравляю с покупкой дома!"
@@ -141,23 +160,24 @@ func callBackResolver(update tg.Update) (tg.MessageConfig, bool) {
 	}
 
 	msg.ParseMode = "markdown"
+	msg.ChatID = update.CallbackQuery.Message.Chat.ID
 	return msg, deletePrevMessage
 }
 
 func useSpecialCell(update tg.Update, char []string, user r.User) tg.MessageConfig {
 	buttons := tg.ReplyKeyboardMarkup{}
-	ItemLeftHand, ItemRightHand, ItemHead := usersHandItemsView(user)
-	msg.ChatID = update.Message.Chat.ID
+	ItemLeftHand, ItemRightHand, ItemHead := helpers.UsersHandItemsView(user)
 
+	// При нажатии кнопок
 	switch char[0] {
 	case v.GetString("message.doing.up"), v.GetString("message.doing.down"), v.GetString("message.doing.left"), v.GetString("message.doing.right"):
-		msg = UserMoving(update, user, char[0])
+		msg = helpers.UserMoving(update, user, char[0])
 	case v.GetString("message.emoji.foot"):
-		msg = UserMoving(update, user, char[1])
+		msg = helpers.UserMoving(update, user, char[1])
 	case v.GetString("message.emoji.hand"), ItemLeftHand.View, ItemRightHand.View:
-		msg = UserUseHandOrInstrument(update, char)
+		msg = helpers.UserUseHandOrInstrument(update, char)
 	case v.GetString("message.emoji.exclamation_mark"):
-		cellLocation := directionCell(update, char[3])
+		cellLocation := helpers.DirectionCell(update, char[3])
 		cell := r.Cell{MapsId: *cellLocation.MapsId, AxisX: *cellLocation.AxisX, AxisY: *cellLocation.AxisY}
 		cell = cell.GetCell()
 		msg.Text = "В зависимости от предмета в твоих руках ты можешь получить разный результат. Выбирай..."
@@ -165,12 +185,10 @@ func useSpecialCell(update tg.Update, char []string, user r.User) tg.MessageConf
 	case v.GetString("message.emoji.stop_use"):
 		msg = tg.NewMessage(update.Message.Chat.ID, v.GetString("errors.user_not_has_item_in_hand"))
 	case "Рюкзак":
-		resUserItems := r.GetBackpackItems(user.ID)
-		msg.Text = MessageBackpackUserItems(resUserItems, 0)
-		msg.ReplyMarkup = helpers.BackpackInlineKeyboard(resUserItems, 0)
+		msg.ReplyMarkup, msg.Text = helpers.BackpackCategoryKeyboard()
 	case "Вещи":
 		userItems := r.GetInventoryItems(user.ID)
-		msg.Text = MessageGoodsUserItems(user, userItems, 0)
+		msg.Text = helpers.MessageGoodsUserItems(user, userItems, 0)
 		msg.ReplyMarkup = helpers.GoodsInlineKeyboard(user, userItems, 0)
 	case v.GetString("message.emoji.online"):
 		userOnline := true
@@ -185,7 +203,7 @@ func useSpecialCell(update tg.Update, char []string, user r.User) tg.MessageConf
 		msg.Text = fmt.Sprintf("%s%sОнлайн выключен!", msg.Text, v.GetString("msg_separator"))
 		msg.ReplyMarkup = buttons
 	case ItemHead.View:
-		res := directionCell(update, char[1])
+		res := helpers.DirectionCell(update, char[1])
 		text, err := r.UpdateUserInstrument(update, user, ItemHead)
 		if err != nil {
 			msg.Text = fmt.Sprintf("%s%s%s", r.ViewItemInfo(res), v.GetString("msg_separator"), text)
@@ -193,26 +211,26 @@ func useSpecialCell(update tg.Update, char []string, user r.User) tg.MessageConf
 			msg.Text = r.ViewItemInfo(res)
 		}
 	case v.GetString("message.emoji.wrench"):
-		loc := directionCell(update, char[1])
+		loc := helpers.DirectionCell(update, char[1])
 		cell := r.Cell{MapsId: *loc.MapsId, AxisX: *loc.AxisX, AxisY: *loc.AxisY}.GetCell()
 		charWorkbench := strings.Fields("workbench usPoint 0 1stComp null 0 2ndComp null 0 3rdComp null 0")
-		msg = Workbench(&cell, charWorkbench)
+		msg = helpers.Workbench(&cell, charWorkbench)
 	case v.GetString("message.emoji.quest"):
-		loc := directionCell(update, char[1])
+		loc := helpers.DirectionCell(update, char[1])
 		cell := r.Cell{MapsId: *loc.MapsId, AxisX: *loc.AxisX, AxisY: *loc.AxisY}.GetCell()
-		msg = Quest(&cell, user)
+		msg = helpers.Quest(&cell, user)
 	default:
 		msg.Text, buttons = r.GetMyMap(update)
 		msg = tg.NewMessage(update.Message.Chat.ID, fmt.Sprintf("%s\n\nНет инструмента в руке!", msg.Text))
 		msg.ReplyMarkup = buttons
 	}
 
+	msg.ChatID = update.Message.Chat.ID
 	return msg
 }
 
 func userMenuLocation(update tg.Update, user r.User) tg.MessageConfig {
 	newMessage := update.Message.Text
-	msg.ChatID = update.Message.Chat.ID
 
 	switch newMessage {
 	case "🗺 Карта 🗺":
@@ -228,6 +246,7 @@ func userMenuLocation(update tg.Update, user r.User) tg.MessageConfig {
 		r.User{MenuLocation: "Меню"}.UpdateUser(update)
 	}
 
+	msg.ChatID = update.Message.Chat.ID
 	return msg
 }
 
@@ -275,13 +294,13 @@ func userProfileLocation(update tg.Update, user r.User) tg.MessageConfig {
 
 func useDefaultCell(update tg.Update, user r.User) tg.MessageConfig {
 	newMessage := update.Message.Text
-	msg.ChatID = update.Message.Chat.ID
 	buttons := tg.ReplyKeyboardMarkup{}
 	currentTime := time.Now()
 
+	// Взаимодействие с предметами на карте, у которых нет действий
 	switch newMessage {
 	case v.GetString("message.doing.up"), v.GetString("message.doing.down"), v.GetString("message.doing.left"), v.GetString("message.doing.right"):
-		msg = UserMoving(update, user, newMessage)
+		msg = helpers.UserMoving(update, user, newMessage)
 	case v.GetString("message.emoji.water"):
 		msg.Text = "Ты не похож на Jesus! 👮‍♂️"
 	case v.GetString("message.emoji.clock"):
@@ -304,570 +323,8 @@ func useDefaultCell(update tg.Update, user r.User) tg.MessageConfig {
 		msg.ReplyMarkup = buttons
 	}
 
+	msg.ChatID = update.Message.Chat.ID
 	return msg
-}
-
-func directionCell(update tg.Update, direction string) r.Location {
-	res := r.GetOrCreateMyLocation(update)
-
-	switch direction {
-	case v.GetString("message.doing.up"):
-		y := *res.AxisY + 1
-		return r.Location{MapsId: res.MapsId, AxisX: res.AxisX, AxisY: &y}
-	case v.GetString("message.doing.down"):
-		y := *res.AxisY - 1
-		return r.Location{MapsId: res.MapsId, AxisX: res.AxisX, AxisY: &y}
-	case v.GetString("message.doing.left"):
-		x := *res.AxisX - 1
-		return r.Location{MapsId: res.MapsId, AxisX: &x, AxisY: res.AxisY}
-	case v.GetString("message.doing.right"):
-		x := *res.AxisX + 1
-		return r.Location{MapsId: res.MapsId, AxisX: &x, AxisY: res.AxisY}
-	}
-	return res
-}
-
-func MessageBackpackUserItems(userItems []r.UserItem, rowUser int) string {
-	var userItemMsg = "🎒 *Рюкзачок*\n \n"
-
-	if len(userItems) == 0 {
-		return userItemMsg + "👻 \U0001F9B4  Пусто .... 🕸 🕷"
-	}
-
-	for i, item := range userItems {
-		var firstCell string
-		switch rowUser {
-		case i:
-			firstCell += item.User.Avatar
-		case i + 1, i - 1:
-			firstCell += "◻️"
-		case i + 2, i - 2:
-			firstCell += "◽️️"
-		default:
-			firstCell += "▫️"
-		}
-		userItemMsg += fmt.Sprintf("%s   %d%s     *HP*:  _%d_ ♥️️     *ST*:  _%d_ \U0001F9C3 ️\n", firstCell, *item.Count, item.Item.View, *item.Item.Healing, *item.Item.Satiety)
-
-	}
-
-	return userItemMsg
-}
-
-func MessageGoodsUserItems(user r.User, userItems []r.UserItem, rowUser int) string {
-	var userItemMsg = "🧥 *Вещички* 🎒\n\n"
-	userItemMsg = messageUserDressedGoods(user) + userItemMsg
-
-	if len(userItems) == 0 {
-		return userItemMsg + "👻 \U0001F9B4  Пусто .... 🕸 🕷"
-	}
-
-	for i, item := range userItems {
-		_, res := user.IsDressedItem(userItems[i])
-
-		if res == v.GetString("callback_char.take_off_good") {
-			res = "✅"
-		} else {
-			res = ""
-		}
-
-		var firstCell string
-		switch rowUser {
-		case i:
-			firstCell += item.User.Avatar
-		case i + 1, i - 1:
-			firstCell += "◻️"
-		case i + 2, i - 2:
-			firstCell += "◽️️"
-		default:
-			firstCell += "▫️"
-		}
-		userItemMsg += fmt.Sprintf("%s  %s %dшт.   %s %s   (%d/%d)\n", firstCell, item.Item.View, *item.Count, res, item.Item.Name, *item.CountUseLeft, *item.Item.CountUse)
-
-	}
-
-	return userItemMsg
-}
-
-func BackPackMoving(charData []string, update tg.Update) tg.MessageConfig {
-	i := r.ToInt(charData[1])
-	userTgId := r.GetUserTgId(update)
-	user := r.GetUser(r.User{TgId: userTgId})
-	userItems := r.GetBackpackItems(user.ID)
-
-	switch i {
-	case len(userItems):
-		i = i - 1
-	}
-
-	msg.Text = MessageBackpackUserItems(userItems, i)
-	msg.ReplyMarkup = helpers.BackpackInlineKeyboard(userItems, i)
-
-	return msg
-}
-
-func GoodsMoving(charData []string, update tg.Update) tg.MessageConfig {
-	i := r.ToInt(charData[1])
-
-	userTgId := r.GetUserTgId(update)
-	user := r.GetUser(r.User{TgId: userTgId})
-	userItems := r.GetInventoryItems(user.ID)
-
-	switch i {
-	case len(userItems):
-		i = i - 1
-	}
-
-	msg.Text = MessageGoodsUserItems(user, userItems, i)
-	msg.ReplyMarkup = helpers.GoodsInlineKeyboard(user, userItems, i)
-
-	return msg
-}
-
-func UserEatItem(update tg.Update, charData []string) tg.MessageConfig {
-	userTgId := r.GetUserTgId(update)
-	userItemId := r.ToInt(charData[1])
-
-	user := r.GetUser(r.User{TgId: userTgId})
-	userItem := r.UserItem{ID: userItemId}.UserGetUserItem()
-
-	res := userItem.EatItem(update, user)
-	charDataForOpenBackPack := strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.backpack_moving"), charData[2]))
-	msg = BackPackMoving(charDataForOpenBackPack, update)
-	msg.Text = fmt.Sprintf("%s%s%s", msg.Text, v.GetString("msg_separator"), res)
-
-	return msg
-}
-
-func UserDeleteItem(update tg.Update, charData []string) tg.MessageConfig {
-	userItemId := r.ToInt(charData[1])
-	userTgId := r.GetUserTgId(update)
-	user := r.GetUser(r.User{TgId: userTgId})
-	userItem := r.UserItem{ID: userItemId}.UserGetUserItem()
-
-	countAfterUserThrowOutItem := 0
-	var updateUserItemStruct = r.UserItem{
-		ID:    userItemId,
-		Count: &countAfterUserThrowOutItem,
-	}
-
-	user.UpdateUserItem(updateUserItemStruct)
-
-	var charDataForOpenList []string
-	switch charData[3] {
-	case "good":
-		charDataForOpenList = strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.goods_moving"), charData[2]))
-		userTakeOffGood(update, charData)
-		msg = GoodsMoving(charDataForOpenList, update)
-	case "backpack":
-		charDataForOpenList = strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.backpack_moving"), charData[2]))
-		msg = BackPackMoving(charDataForOpenList, update)
-	}
-
-	msg.Text = fmt.Sprintf("%s%s🗑 Вы выкинули %s%dшт.", msg.Text, v.GetString("msg_separator"), userItem.Item.View, *userItem.Count)
-
-	return msg
-}
-
-func usersHandItemsView(user r.User) (r.Item, r.Item, r.Item) {
-	ItemLeftHand := r.Item{View: v.GetString("message.emoji.hand")}
-	ItemRightHand := r.Item{View: v.GetString("message.emoji.hand")}
-	var ItemHead r.Item
-
-	if user.LeftHand != nil {
-		ItemLeftHand = *user.LeftHand
-	}
-	if user.RightHand != nil {
-		ItemRightHand = *user.RightHand
-	}
-	if user.Head != nil {
-		ItemHead = *user.Head
-	}
-
-	return ItemLeftHand, ItemRightHand, ItemHead
-}
-
-func messageUserDressedGoods(user r.User) string {
-	var head string
-	var body string
-	var leftHand string
-	var rightHand string
-	var foot string
-	var shoes string
-
-	if user.Head != nil {
-		head = user.Head.View
-	} else {
-		head = "🟦"
-	}
-	if user.LeftHand != nil {
-		leftHand = user.LeftHand.View
-	} else {
-		leftHand = "✋"
-	}
-	if user.RightHand != nil {
-		rightHand = user.RightHand.View
-	} else {
-		rightHand = "🤚"
-	}
-	if user.Body != nil {
-		body = user.Body.View
-	} else {
-		body = "👔"
-	}
-	if user.Foot != nil {
-		foot = user.Foot.View
-	} else {
-		foot = "\U0001FA72"
-	}
-	if user.Shoes != nil {
-		shoes = user.Shoes.View
-	} else {
-		shoes = "👣"
-	}
-
-	var messageUserGoods = fmt.Sprintf("〰️☁️〰️〰️☀️\n"+
-		"〰️〰️%s〰️〰️\n"+
-		"〰️〰️%s〰️〰️\n"+
-		"〰️%s%s%s〰️\n"+
-		"〰️〰️%s〰️〰️\n"+
-		"〰️〰️%s〰️️🍺️\n"+
-		"\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\n\n",
-		head, user.Avatar, leftHand, body, rightHand, foot, shoes)
-
-	return messageUserGoods
-}
-
-func userTakeOffGood(update tg.Update, charData []string) {
-	userItemId := r.ToInt(charData[1])
-	userTgId := r.GetUserTgId(update)
-	user := r.GetUser(r.User{TgId: userTgId})
-	userItem := r.UserItem{ID: userItemId}.UserGetUserItem()
-
-	if user.HeadId != nil && userItem.ItemId == *user.HeadId {
-		r.SetNullUserField(update, "head_id")
-	} else if user.LeftHandId != nil && userItem.ItemId == *user.LeftHandId {
-		r.SetNullUserField(update, "left_hand_id")
-	} else if user.RightHandId != nil && userItem.ItemId == *user.RightHandId {
-		r.SetNullUserField(update, "right_hand_id")
-	} else if user.BodyId != nil && userItem.ItemId == *user.BodyId {
-		r.SetNullUserField(update, "body_id")
-	} else if user.FootId != nil && userItem.ItemId == *user.FootId {
-		r.SetNullUserField(update, "foot_id")
-	} else if user.ShoesId != nil && userItem.ItemId == *user.ShoesId {
-		r.SetNullUserField(update, "shoes_id")
-	}
-
-	charDataForOpenGoods := strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.goods_moving"), charData[2]))
-	msg = GoodsMoving(charDataForOpenGoods, update)
-	msg.Text = fmt.Sprintf("%s%sВещь снята!", msg.Text, v.GetString("msg_separator"))
-}
-
-func dressUserItem(update tg.Update, charData []string) tg.MessageConfig {
-	userItemId := r.ToInt(charData[1])
-	userTgId := r.GetUserTgId(update)
-	user := r.GetUser(r.User{TgId: userTgId})
-	userItem := r.UserItem{ID: userItemId}.UserGetUserItem()
-	changeHandItem := false
-
-	var result = fmt.Sprintf("Вы надели %s", userItem.Item.View)
-
-	switch *userItem.Item.DressType {
-	case "hand":
-		if user.LeftHandId == nil {
-			r.User{LeftHandId: &userItem.ItemId}.UpdateUser(update)
-		} else if user.RightHandId == nil {
-			r.User{RightHandId: &userItem.ItemId}.UpdateUser(update)
-		} else {
-			result = "У вас заняты все руки! Что хочешь снять?"
-			changeHandItem = true
-		}
-	case "head":
-		if user.HeadId == nil {
-			r.User{HeadId: &userItem.ItemId}.UpdateUser(update)
-		} else {
-			result = v.GetString("errors.user_has_other_item")
-		}
-	case "body":
-		if user.BodyId == nil {
-			r.User{BodyId: &userItem.ItemId}.UpdateUser(update)
-		} else {
-			result = v.GetString("errors.user_has_other_item")
-		}
-	case "foot":
-		if user.FootId == nil {
-			r.User{FootId: &userItem.ItemId}.UpdateUser(update)
-		} else {
-			result = v.GetString("errors.user_has_other_item")
-		}
-	case "shoes":
-		if user.ShoesId == nil {
-			r.User{ShoesId: &userItem.ItemId}.UpdateUser(update)
-		} else {
-			result = v.GetString("errors.user_has_other_item")
-		}
-	}
-
-	if changeHandItem {
-		msg.ReplyMarkup = helpers.ChangeItemInHand(user, userItemId, charData[2])
-	} else {
-		charDataForOpenGoods := strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.goods_moving"), charData[2]))
-		msg = GoodsMoving(charDataForOpenGoods, update)
-	}
-
-	msg.Text = fmt.Sprintf("%s%s%s", msg.Text, v.GetString("msg_separator"), result)
-
-	return msg
-}
-
-func userThrowOutItem(update tg.Update, user r.User, charData []string) tg.MessageConfig {
-	userItem := r.UserItem{ID: r.ToInt(charData[1])}.UserGetUserItem()
-
-	*userItem.Count = *userItem.Count - r.ToInt(charData[3])
-
-	err := r.UpdateCellUnderUser(update, userItem, r.ToInt(charData[3]))
-	var msgtext string
-	if err != nil {
-		msgtext = fmt.Sprintf("%s%s", v.GetString("msg_separator"), err)
-	} else {
-		msgtext = fmt.Sprintf("%sВы сбросили %s %sшт. на карту!", v.GetString("msg_separator"), userItem.Item.View, charData[3])
-		user.UpdateUserItem(r.UserItem{ID: userItem.ID, Count: userItem.Count})
-	}
-
-	var charDataForOpenList []string
-	switch charData[4] {
-	case "good":
-		charDataForOpenList = strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.goods_moving"), charData[2]))
-		if *userItem.Count == 0 {
-			userTakeOffGood(update, charData)
-		}
-		msg = GoodsMoving(charDataForOpenList, update)
-	case "backpack":
-		charDataForOpenList = strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.backpack_moving"), charData[2]))
-		msg = BackPackMoving(charDataForOpenList, update)
-	}
-
-	msg.Text = fmt.Sprintf("%s%s", msg.Text, msgtext)
-
-	return msg
-}
-
-func userWantsToThrowOutItem(update tg.Update, charData []string) tg.MessageConfig {
-	userItem := r.UserItem{ID: r.ToInt(charData[1])}.UserGetUserItem()
-
-	if userItem.CountUseLeft != nil && *userItem.CountUseLeft != *userItem.Item.CountUse {
-		*userItem.Count = *userItem.Count - 1
-	}
-
-	if *userItem.Count == 0 {
-		var charDataForOpenList []string
-		switch charData[3] {
-		case "good":
-			charDataForOpenList = strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.goods_moving"), charData[2]))
-			if *userItem.CountUseLeft == *userItem.Item.CountUse {
-				userTakeOffGood(update, charData)
-			}
-			msg = GoodsMoving(charDataForOpenList, update)
-		case "backpack":
-			charDataForOpenList = strings.Fields(fmt.Sprintf("%s %s", v.GetString("callback_char.backpack_moving"), charData[2]))
-			msg = BackPackMoving(charDataForOpenList, update)
-		}
-		msg.Text = fmt.Sprintf("%s%sНельзя выкинуть на карту предмет, который уже был использован!", msg.Text, v.GetString("msg_separator"))
-	} else {
-		msg.ReplyMarkup = helpers.CountItemUserWantsToThrow(charData, userItem)
-		msg.Text = fmt.Sprintf("%sСколько %s ты хочешь скинуть?", v.GetString("msg_separator"), userItem.Item.View)
-	}
-
-	return msg
-}
-
-func Quest(cell *r.Cell, user r.User) tg.MessageConfig {
-	if !cell.IsQuest() {
-		msg.Text = v.GetString("error.no_quest_item")
-		return msg
-	}
-
-	msg.Text = v.GetString("user_location.tasks_menu_message")
-	msg.ReplyMarkup = helpers.AllQuestsMessageKeyboard(user)
-
-	return msg
-}
-
-func Workbench(cell *r.Cell, char []string) tg.MessageConfig {
-	var charData []string
-	if !cell.IsWorkbench() {
-		msg.Text = "Здесь нет верстака!"
-		return msg
-	}
-
-	if cell != nil {
-		charData = strings.Fields("workbench usPoint 0 1stComp nil 0 2ndComp nil 0 3rdComp nil 0")
-	} else {
-		charData = strings.Fields(fmt.Sprintf("workbench usPoint %s 1stComp %s %s 2ndComp %s %s 3rdComp %s %s", char[2], char[4], char[5], char[7], char[8], char[10], char[11]))
-	}
-
-	msg = OpenWorkbenchMessage(charData)
-	msg.ReplyMarkup = helpers.WorkbenchButton(charData)
-
-	return msg
-}
-
-func OpenWorkbenchMessage(char []string) tg.MessageConfig {
-	// char = workbench usPoint 0 1stComp: id 0 2ndComp id 0 3rdComp id 0
-
-	fstCnt := getViewEmojiForMsg(char, 0)
-	secCnt := getViewEmojiForMsg(char, 1)
-	trdCnt := getViewEmojiForMsg(char, 2)
-
-	fstComponentView := viewComponent(char[4])
-	secComponentView := viewComponent(char[7])
-	trdComponentView := viewComponent(char[10])
-
-	cellUser := r.ToInt(char[2])
-	userPointer := strings.Fields("〰️ 〰️ 〰️")
-	userPointer[cellUser] = "👇"
-
-	msg.Text = fmt.Sprintf(
-		"〰️〰️〰️☁️〰️〰️〰️☀️〰️\n"+
-			"〰️〰️%s〰️%s〰️%s〰️〰️\n"+
-			"🔬〰️%s〰️%s〰️%s〰️📡\n"+
-			"\U0001F7EB\U0001F7EB%s\U0001F7EB%s\U0001F7EB%s\U0001F7EB\U0001F7EB\n"+
-			"〰️\U0001F7EB〰️〰️〰️〰️〰️\U0001F7EB〰️\n"+
-			"〰️\U0001F7EB〰️〰️〰️🍺〰️\U0001F7EB〰️\n"+
-			"\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9\U0001F7E9",
-		userPointer[0], userPointer[1], userPointer[2],
-		fstComponentView, secComponentView, trdComponentView,
-		fstCnt, secCnt, trdCnt)
-
-	return msg
-}
-
-func viewComponent(id string) string {
-	if id != "nil" {
-		component := r.UserItem{ID: r.ToInt(id)}.UserGetUserItem()
-		return component.Item.View
-	}
-	return "⚪"
-}
-
-func AllReceiptsMsg() string {
-	receipts := r.GetReceipts()
-	var msgText string
-	for _, receipt := range receipts {
-		var fstEl string
-		var secEl string
-		var trdEl string
-
-		if receipt.Component1ID != nil {
-			fstEl = fmt.Sprintf("%d⃣%s", *receipt.Component1Count, receipt.Component1.View)
-		}
-		if receipt.Component2ID != nil {
-			secEl = fmt.Sprintf("➕%d⃣%s", *receipt.Component2Count, receipt.Component2.View)
-		}
-		if receipt.Component3ID != nil {
-			trdEl = fmt.Sprintf("➕%d⃣%s", *receipt.Component3Count, receipt.Component3.View)
-		}
-		msgText = msgText + fmt.Sprintf("%s 🔚 %s%s%s\n", receipt.ItemResult.View, fstEl, secEl, trdEl)
-	}
-	return msgText
-}
-
-func GetReceiptFromData(char []string) r.Receipt {
-	var result r.Receipt
-
-	if char[4] != "nil" && char[5] != "0" {
-		fstItem := r.UserItem{ID: r.ToInt(char[4])}.UserGetUserItem()
-		id := int(fstItem.Item.ID)
-		c := r.ToInt(char[5])
-
-		result.Component1ID = &id
-		result.Component1Count = &c
-	}
-
-	if char[7] != "nil" && char[8] != "0" {
-		fstItem := r.UserItem{ID: r.ToInt(char[7])}.UserGetUserItem()
-		id := int(fstItem.Item.ID)
-		c := r.ToInt(char[8])
-
-		result.Component2ID = &id
-		result.Component2Count = &c
-	}
-
-	if char[10] != "nil" && char[11] != "0" {
-		fstItem := r.UserItem{ID: r.ToInt(char[10])}.UserGetUserItem()
-		id := int(fstItem.Item.ID)
-		c := r.ToInt(char[11])
-
-		result.Component3ID = &id
-		result.Component3Count = &c
-	}
-
-	return result
-}
-
-func PutCountComponent(char []string) tg.MessageConfig {
-	userItemId := char[r.ToInt(char[2])+(4+r.ToInt(char[2])*2)] // char[x + (4+x*2 )] = char[4]
-	userItem := r.UserItem{ID: r.ToInt(userItemId)}.UserGetUserItem()
-
-	msg.ReplyMarkup = helpers.ChangeCountUserItem(char, userItem)
-
-	return msg
-}
-
-func UserCraftItem(user r.User, receipt *r.Receipt) (tg.MessageConfig, bool) {
-	deletePrevMessage := true
-	if receipt == nil {
-		msg.Text = "Такого рецепта не существует!"
-		msg.ReplyMarkup = nil
-		deletePrevMessage = false
-		return msg, deletePrevMessage
-	}
-
-	msg.ReplyMarkup = nil
-	resultItem := r.UserItem{UserId: int(user.ID), ItemId: receipt.ItemResultID}.UserGetUserItem()
-
-	if *receipt.ItemResultCount+*resultItem.Count > *resultItem.Item.MaxCountUserHas {
-		msg.Text = fmt.Sprintf("Вы не можете иметь больше, чем %d %s!\nСори... такие правила(", *resultItem.Item.MaxCountUserHas, resultItem.Item.View)
-		msg.ReplyMarkup = nil
-		deletePrevMessage = false
-		return msg, deletePrevMessage
-	}
-
-	if receipt.Component1ID != nil && receipt.Component1Count != nil {
-		userItem := r.UserItem{UserId: int(user.ID), ItemId: *receipt.Component1ID}.UserGetUserItem()
-		countItem1 := *userItem.Count - *receipt.Component1Count
-		user.UpdateUserItem(r.UserItem{ID: userItem.ID, ItemId: *receipt.Component1ID, Count: &countItem1}) // CountUseLeft: resultItem.CountUseLeft
-	}
-	if receipt.Component2ID != nil && receipt.Component2Count != nil {
-		userItem := r.UserItem{UserId: int(user.ID), ItemId: *receipt.Component2ID}.UserGetUserItem()
-		countItem2 := *userItem.Count - *receipt.Component2Count
-		user.UpdateUserItem(r.UserItem{ID: userItem.ID, ItemId: *receipt.Component2ID, Count: &countItem2}) // CountUseLeft: resultItem.CountUseLeft
-	}
-	if receipt.Component3ID != nil && receipt.Component3Count != nil {
-		userItem := r.UserItem{UserId: int(user.ID), ItemId: *receipt.Component3ID}.UserGetUserItem()
-		countItem3 := *userItem.Count - *receipt.Component3Count
-		user.UpdateUserItem(r.UserItem{ID: userItem.ID, ItemId: *receipt.Component3ID, Count: &countItem3}) // CountUseLeft: resultItem.CountUseLeft
-	}
-
-	if *resultItem.Count == 0 {
-		resultItem.CountUseLeft = resultItem.Item.CountUse
-	}
-	*resultItem.Count = *resultItem.Count + *receipt.ItemResultCount
-	user.UpdateUserItem(r.UserItem{ID: resultItem.ID, Count: resultItem.Count, CountUseLeft: resultItem.CountUseLeft})
-
-	charData := strings.Fields("workbench usPoint 0 1stComp nil 0 2ndComp nil 0 3rdComp nil 0")
-	msg = Workbench(nil, charData)
-	msg.Text = fmt.Sprintf("%s%sСупер! Ты получил %s %d шт. %s!", msg.Text, v.GetString("msg_separator"), resultItem.Item.View, *receipt.ItemResultCount, receipt.ItemResult.Name)
-	return msg, deletePrevMessage
-}
-
-func getViewEmojiForMsg(char []string, i int) string {
-	count := i + 5 + i*2
-
-	if char[count] == "0" {
-		return "\U0001F7EB"
-	}
-
-	return fmt.Sprintf("%s⃣", char[count])
 }
 
 func updateUserHand(update tg.Update, char []string, userItem r.UserItem) {
@@ -877,71 +334,4 @@ func updateUserHand(update tg.Update, char []string, userItem r.UserItem) {
 	case v.GetString("callback_char.change_right_hand"):
 		r.User{RightHandId: &userItem.ItemId}.UpdateUser(update)
 	}
-}
-
-func UserMoving(update tg.Update, user r.User, char string) tg.MessageConfig {
-	var text string
-	res := directionCell(update, char)
-
-	var inlineButtons tg.InlineKeyboardMarkup
-
-	locMsg, err := r.UpdateLocation(update, res, user)
-	msgMap, buttons := r.GetMyMap(update)
-
-	if err != nil {
-		if err.Error() == "user has not home" {
-			inlineButtons = helpers.BuyHomeKeyboard()
-			msg.ReplyMarkup = inlineButtons
-			text = locMsg
-		} else {
-			text = fmt.Sprintf("%s%s%s", msgMap, v.GetString("msg_separator"), locMsg)
-		}
-	} else {
-		msg.ReplyMarkup = buttons
-
-		lighterMsg, err := user.CheckUserHasLighter(update)
-		if err != nil {
-			text = fmt.Sprintf("%s%s", v.GetString("msg_separator"), lighterMsg)
-		}
-		text = fmt.Sprintf("%s%s", msgMap, text)
-	}
-
-	msg.Text = text
-	return msg
-}
-
-func UserUseHandOrInstrument(update tg.Update, char []string) tg.MessageConfig {
-	res := directionCell(update, char[1])
-	resultOfGetItem := r.UserGetItem(update, res, char)
-	resText, buttons := r.GetMyMap(update)
-	msg.Text = fmt.Sprintf("%s%s%s", resText, v.GetString("msg_separator"), resultOfGetItem)
-	msg.ReplyMarkup = buttons
-
-	return msg
-}
-
-func OpenQuest(questId uint, user r.User) tg.MessageConfig {
-	quest := r.Quest{ID: questId}.GetQuest()
-	userQuest := r.UserQuest{UserId: user.ID, QuestId: questId}.GetUserQuest()
-
-	msg.Text = quest.QuestInfo(userQuest)
-	msg.ReplyMarkup = helpers.OpenQuestKeyboard(quest, userQuest)
-
-	return msg
-}
-
-func UserDoneQuest(questId uint, user r.User) tg.MessageConfig {
-	userQuest := r.UserQuest{UserId: user.ID, QuestId: questId}.GetUserQuest()
-	if !userQuest.Quest.Task.HasUserDoneTask(user) {
-		msg.Text = v.GetString("errors.user_did_not_task")
-		return msg
-	}
-
-	userQuest.UserDoneQuest(user)
-	user.UserGetResult(userQuest.Quest.Result)
-
-	msg = OpenQuest(questId, user)
-	msg.Text = fmt.Sprintf("*Задание выполнено!*\n%s", msg.Text)
-
-	return msg
 }
